wip, creating uuid generator func (employee)
