+++ c/employee-manager/Makefile
@@ -1,0 +1,2 @@
+compile:
+	gcc -o bin/employee-manager src/main.c src/classes/employee.c -Iheaders/
